ATR Core Russian Translation courtesy of LeonusDH
* Russian Language (Part 1)

The first part of the translated text. I'm just committing the changes.

God Almighty ...
That's an awful lot of files...

* Russian Language (Part 2)

The last part of the translation.

In general, everything is as usual. AI and personal skills for translation. 98% confidence in the accuracy of the translation.

P.S. Don't add new lines for translation so soon, please 😥